7. SEO_Opportunities_Finder/requirements.txt
pandas>=2.1.4,<2.3.0
numpy==1.26.2
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0
python-dotenv>=1.0.0,<2.0.2
tqdm==4.66.1
setuptools
wheel